pyproject.toml
[project]
name = "mcp-sqlalchemy"
version = "0.1.0"
description = "MCP server for SQLAlchemy database interactions - supports SQLite, PostgreSQL, and MySQL"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Tomer Gabay", email = "tomer.gabay@woonstadrotterdam.nl"},
    {name = "Ben Verhees", email = "ben.verhees@woonstadrotterdam.nl"},
]
keywords = ["mcp", "sqlalchemy", "database", "postgresql", "mysql", "sqlite"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "aiosqlite>=0.21.0",
    "greenlet>=3.0.0",
    "mcp[cli]>=1.12.2",
    "sqlalchemy>=2.0.42",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
postgresql = ["asyncpg>=0.28.0"]
mysql = ["aiomysql>=0.1.1"]
all = ["asyncpg>=0.28.0", "aiomysql>=0.1.1"]

[project.scripts]
mcp-sqlalchemy= "mcp_sqlalchemy.__main__:main"

[build-system]
requires = ["setuptools"]
build-backed = "setuptools.build_meta"

[tools.setuptools]
package_dir = {"" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
